+++ prp/io/manifest.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 import yaml
 
-from prp.models.manifest import SampleManifest
+from prp.models.manifest import SampleManifest, BootstrapConfig
 
 from .types import Pathish
 
@@ -18,6 +18,14 @@ def read_manifest(path: Pathish) -> SampleManifest:
     if not path.is_file():
         raise FileNotFoundError(f"file {path.name} not found, please check the path.")
 
-    with path.open(encoding="utf-8") as cfile:
-        data = yaml.safe_load(cfile)
+    with path.open(encoding="utf-8") as f:
+        data = yaml.safe_load(f)
         return SampleManifest.model_validate(data, context=path)
+
+
+def read_bootstrap_config(path: Pathish) -> BootstrapConfig:
+    """Read boostrap configuration."""
+    path = Path(path)
+    with path.open(encoding="utf-8") as f:
+        data = yaml.safe_load(f)
+        return BootstrapConfig.model_validate(data, context=path)