+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 
+- Added methods for ensuring existance of users and groups.
 - Added a result model registry and function for casting JSON results to a result data model.
 - Added APIs for registering and excecuting parses from the input software name and version.
 - Added parser registry with support for version based parser selection.